requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sentence-transformers
faiss-cpu
langchain
openai